% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samuika.R
\encoding{UTF-8}
\name{boot_samuika}
\alias{boot_samuika}
\title{Bootstrap samuika}
\usage{
boot_samuika(model, n = 100, seed = 1, bias_correct = model$input$bias_correct)
}
\arguments{
\item{model}{\code{samuika} object}

\item{n}{the number of bootstrap replicates}

\item{seed}{seed for \code{set.seed}}

\item{bias_correct}{bias correct option in \code{sdreport}}
}
\description{
Bootstrap samuika
}
