% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlitika.R
\encoding{UTF-8}
\name{compare_steady_state}
\alias{compare_steady_state}
\title{Plot to compare steady states against different F values}
\usage{
compare_steady_state(trace_list, trace_name = NULL, base_size = 16, Fmax = 1.5)
}
\arguments{
\item{trace_list}{list of (multiple ) \code{trace_future} objects}
}
\description{
Plot to compare steady states against different F values
}
