% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlitika.R
\encoding{UTF-8}
\name{convert_samuika_estMSY}
\alias{convert_samuika_estMSY}
\title{Function for converting samuika result to the data for est_MSY}
\usage{
convert_samuika_estMSY(samuika_res, Stock_ID = 0, scale_num_to_mass = NULL)
}
\arguments{
\item{samuika_res}{\code{samuika} object}

\item{Stock_ID}{Stock_ID (default is 0)}
}
\description{
Function for converting samuika result to the data for est_MSY
}
