% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messir.R
\encoding{UTF-8}
\name{det_BRF}
\alias{det_BRF}
\title{Calculating deterministic biological reference points}
\usage{
det_BRF(rec_pars)
}
\arguments{
\item{rec_pars}{\code{get_rec_pars} object}
}
\description{
Calculating deterministic biological reference points
}
