% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messir.R
\encoding{UTF-8}
\name{est_MSY}
\alias{est_MSY}
\title{Estimating MSY-based management reference points for Surume-Ika}
\usage{
est_MSY(
  future_sim_res,
  obj_catch = "mean",
  nsim = NULL,
  MSY_year = NULL,
  PGY = NULL,
  PGYlower_only = FALSE,
  percentB0 = NULL,
  Bempirical = NULL,
  seed = NULL,
  TMB = FALSE,
  nlminb_control = list(eval.max = 2000, iter.max = 2000)
)
}
\arguments{
\item{future_sim_res}{\code{future_sim} object}

\item{obj_catch}{objective of maximum catch: mean(0) or geomean(1)}
}
\description{
Estimating MSY-based management reference points for Surume-Ika
}
