% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messir.R
\encoding{UTF-8}
\name{future_sim}
\alias{future_sim}
\title{future simulation for Surume-Ika}
\usage{
future_sim(
  assess_data,
  Fcurrent,
  multi = 1,
  sim_year = 50,
  nsim = 1000,
  Fcurrent_year = NULL,
  SR = "HS",
  rec_arg = list(a = NULL, b = NULL, sd = 0, rho = 0, resid = NULL),
  bias_correct = TRUE,
  weight,
  num_to_mass_scale = 1/1000,
  M = 0.6,
  Pope = TRUE,
  seed = 12345,
  HCR = FALSE,
  SBrefs = c("SBlimit", "SBban"),
  Ftarget = NULL,
  array_to_tibble = FALSE,
  sim_rec_resid = NULL,
  scenario_name = NULL
)
}
\arguments{
\item{assess_data}{assessment data}

\item{Fcurrent}{current fishing mortality coefficient}
}
\description{
future simulation for Surume-Ika
}
