% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samuika.R
\encoding{UTF-8}
\name{get_mohn}
\alias{get_mohn}
\title{Calculation of Mohn's rho from retrospective analysis of samuika}
\usage{
get_mohn(retro_res, first_eval_year = NULL, lag_year = 1)
}
\arguments{
\item{retro_res}{retro_samuika object}

\item{first_eval_year}{firstly-evaluated year}

\item{lag_year}{time-lag in evaluation of Mohn's rho}
}
\description{
Calculation of Mohn's rho from retrospective analysis of samuika
}
