% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlitika.R
\encoding{UTF-8}
\name{get_prob_ref}
\alias{get_prob_ref}
\title{Function for calculating probability of achiveing biological reference points}
\usage{
get_prob_ref(future_sim_res, ref_value, variable = "Spawning_biomass")
}
\arguments{
\item{future_sim_res}{\code{future_sim} object}

\item{ref_value}{value of biological reference points}

\item{variable}{what variable ref_value indicates (default: "Spawning_biomass")}
}
\description{
Function for calculating probability of achiveing biological reference points
}
