% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samuika.R
\encoding{UTF-8}
\name{get_rec_pars}
\alias{get_rec_pars}
\title{Function for deriving stock-recruitment parameters}
\usage{
get_rec_pars(model, M = model$input$M)
}
\arguments{
\item{model}{samuika object}
}
\description{
Function for deriving stock-recruitment parameters
}
