% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlitika.R
\encoding{UTF-8}
\name{get_SRdata}
\alias{get_SRdata}
\title{Making a combined list with original list name(s)}
\usage{
get_SRdata(model)
}
\arguments{
\item{model}{samuika object}
}
\description{
Making a combined list with original list name(s)
}
