% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samuika.R
\encoding{UTF-8}
\name{integrate_detBRF}
\alias{integrate_detBRF}
\title{Output summary table with deterministic biological reference points}
\usage{
integrate_detBRF(model, M = model$input$M)
}
\arguments{
\item{model}{\code{samuika} object}
}
\description{
Output summary table with deterministic biological reference points
}
