% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlitika.R
\encoding{UTF-8}
\name{make_named_list}
\alias{make_named_list}
\title{Making a combined list with original list name(s)}
\usage{
make_named_list(...)
}
\arguments{
\item{...}{list of \code{class("list")}}
}
\description{
Making a combined list with original list name(s)
}
