% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlitika.R
\encoding{UTF-8}
\name{out_beta_table}
\alias{out_beta_table}
\title{output summary table when changing beta}
\usage{
out_beta_table(
  future_HCR,
  SBmsy,
  Fmsy = NULL,
  beta = seq(0, 1, by = 0.1),
  future_Fcurrent = NULL,
  filename = "beta_table"
)
}
\arguments{
\item{future_HCR}{\code{future_sim} object with \code{HCR = TRUE}}

\item{SBmsy}{SBmsy as a target reference point}

\item{Fmsy}{Fmsy to be multiplied by beta}

\item{beta}{vector of beta (multipliers of Fmsy)}

\item{future_Fcurrent}{\code{future_sim} object with Fcurrent (optional)}

\item{filename}{file name of saving}
}
\description{
output summary table when changing beta
}
