% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlitika.R
\encoding{UTF-8}
\name{out_boot_table}
\alias{out_boot_table}
\title{Output summary table for bootstrapped samuika}
\usage{
out_boot_table(boot_res, CI_range = 0.8)
}
\arguments{
\item{boot_res}{\code{boot_samuika} object}
}
\description{
Output summary table for bootstrapped samuika
}
