% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlitika.R
\encoding{UTF-8}
\name{out_Fmulti_table}
\alias{out_Fmulti_table}
\title{output summary table when changing Fmulti}
\usage{
out_Fmulti_table(
  future_Fcurrent,
  SBrefs,
  Fmulti = seq(0, 1, by = 0.1),
  scenario_name = NULL,
  filename = "Fmulti_table"
)
}
\arguments{
\item{future_Fcurrent}{\code{future_sim} object with Fcurrent}

\item{SBrefs}{Reference points for calculating achievement probability}

\item{filename}{file name of saving}
}
\description{
output summary table when changing Fmulti
}
