% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlitika.R
\encoding{UTF-8}
\name{out_summary_estimate}
\alias{out_summary_estimate}
\title{output a table of samuika results}
\usage{
out_summary_estimate(model)
}
\arguments{
\item{model}{\code{samuika} object}
}
\description{
output a table of samuika results
}
