% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlitika.R
\encoding{UTF-8}
\name{out_summary_estimate_n}
\alias{out_summary_estimate_n}
\title{output a table of samuika results (Stock number and Spawning_number)}
\usage{
out_summary_estimate_n(model, CI = 0.8)
}
\arguments{
\item{model}{\code{samuika} object}
}
\description{
output a table of samuika results (Stock number and Spawning_number)
}
