% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlitika.R
\encoding{UTF-8}
\name{plot_boot_F}
\alias{plot_boot_F}
\title{plot of bootstrapped fishing mortality coefficient}
\usage{
plot_boot_F(
  res,
  boot_res,
  CI = 0.8,
  Stock_name = NULL,
  alpha = 0.3,
  path_size = 1.5,
  colour = "red",
  ...
)
}
\arguments{
\item{res}{\code{samuika} object}

\item{boot_res}{\code{boot_samuika} object}
}
\description{
plot of bootstrapped fishing mortality coefficient
}
