% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlitika.R
\encoding{UTF-8}
\name{plot_Catch_HCR}
\alias{plot_Catch_HCR}
\title{Plot Catch biomass against SSB under HCR}
\usage{
plot_Catch_HCR(
  SBtarget,
  future_HCR,
  length = 1000,
  col.multi2currf = "black",
  col.SBtarget = "#00533E",
  col.SBlim = "#edb918",
  col.SBban = "#C73C2E",
  col.Ftarget = "black",
  col.betaFtarget = "gray",
  is.text = TRUE,
  RP.label = c("目標管理基準値案", "限界管理基準値案",
    "禁漁水準案")
)
}
\arguments{
\item{SBtarget}{value of SBtarget}

\item{future_HCR}{\code{future_sim} object with \code{HCR = TRUE}}
}
\description{
Plot Catch biomass against SSB under HCR
}
