% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlitika.R
\encoding{UTF-8}
\name{plot_retro_samuika}
\alias{plot_retro_samuika}
\title{plotting retrospective analyses of samuika}
\usage{
plot_retro_samuika(
  samuika_retro,
  mohn = NULL,
  Stock_ID = 0,
  latest_year = NULL,
  max_retro_year = NULL,
  forecast_year = 0
)
}
\arguments{
\item{samuika_retro}{samuika_retro object}
}
\description{
plotting retrospective analyses of samuika
}
