% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlitika.R
\encoding{UTF-8}
\name{plot_samuika_estimates}
\alias{plot_samuika_estimates}
\title{plotting (multiple) samuika results}
\usage{
plot_samuika_estimates(
  list_samuika_res,
  model_name = NULL,
  Stock_ID = 0,
  CI = 0.8,
  CI_plot = TRUE,
  legend_position = "bottom",
  legend_text_size = 1.2,
  legend_key_size = 1
)
}
\arguments{
\item{list_samuika_res}{list of samuika objects}
}
\description{
plotting (multiple) samuika results
}
