% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlitika.R
\encoding{UTF-8}
\name{plot_SR_regime}
\alias{plot_SR_regime}
\title{plotting stock-recruitment curve with regimes from a samuika object}
\usage{
plot_SR_regime(
  list_samuika_res,
  model_name = NULL,
  Stock_ID = 0,
  regime_name = NULL,
  plot_point = TRUE,
  draw_replace = TRUE,
  Smsy = NULL,
  xmax_ratio = 1.3,
  ymax_ratio = 1.1,
  path_size = 1.2,
  legend_position = "right",
  legend_text_size = 1.2,
  legend_key_size = 1,
  legend_off = TRUE,
  palette_name = "Set1",
  point_size = 2,
  connect_point = TRUE,
  connect_point_size = 1,
  label_year = NULL
)
}
\arguments{
\item{list_samuika_res}{list of samuika objects}

\item{model_name}{model names used in legend}

\item{regime_name}{regime names used in legend}
}
\description{
plotting stock-recruitment curve with regimes from a samuika object
}
