% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlitika.R
\encoding{UTF-8}
\name{plot_SR_samuika}
\alias{plot_SR_samuika}
\title{plotting stock-recruitment curve from samuika objects}
\usage{
plot_SR_samuika(
  list_samuika_res,
  model_name = NULL,
  Stock_ID = 0,
  plot_point = TRUE,
  draw_replace = TRUE,
  Smsy = NULL,
  xmax_ratio = 1.3,
  ymax_ratio = 1.1,
  path_size = 1.2,
  legend_position = "right",
  legend_text_size = 1.2,
  legend_key_size = 1
)
}
\arguments{
\item{list_samuika_res}{list of samuika objects}

\item{model_name}{model names used in legend}
}
\description{
plotting stock-recruitment curve from samuika objects
}
