% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samuika.R
\encoding{UTF-8}
\name{retro_samuika}
\alias{retro_samuika}
\title{Retrospective analysis for samuika}
\usage{
retro_samuika(samuika_res, n = 5)
}
\arguments{
\item{samuika_res}{samuika object}

\item{n}{number of maximum removed years}
}
\description{
Retrospective analysis for samuika
}
