% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messir.R
\encoding{UTF-8}
\name{simulate_rec_resid}
\alias{simulate_rec_resid}
\title{Function for simulating recruitment residuals with consideration for bias correction (if neccessary)}
\usage{
simulate_rec_resid(
  sd,
  rho = 0,
  resample = FALSE,
  resampled_resid = NULL,
  bias_correct = TRUE,
  resid_for_bias_correction = NULL,
  year = NULL,
  nsim = NULL,
  seed = 1,
  out = "resid"
)
}
\description{
Function for simulating recruitment residuals with consideration for bias correction (if neccessary)
}
