% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messir.R
\encoding{UTF-8}
\name{trace_future}
\alias{trace_future}
\title{Tracing for future_sim results with different F values}
\usage{
trace_future(
  future_sim_res,
  trace_multi = seq(0, 10, by = 0.1),
  obj_catch = "mean",
  sim_year = NULL,
  nsim = NULL,
  seed = NULL
)
}
\arguments{
\item{future_sim_res}{\code{future_sim} object}

\item{trace_multi}{multiplier for Fcurrent to be traced}

\item{obj_catch}{objective value("mean", "median, or "geomean")}

\item{sim_year}{simulation years}

\item{nsim}{number of simulations}

\item{seed}{seed in \code{set.seed()}}
}
\description{
Tracing for future_sim results with different F values
}
