% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_cpp.R
\encoding{UTF-8}
\name{use_est_MSY_tmb}
\alias{use_est_MSY_tmb}
\title{Function for activating cpp file for samuuika with TMB}
\usage{
use_est_MSY_tmb(
  TmbFile = "est_MSY_tmb",
  CppDir = system.file("inst/executable", package = "messir"),
  RunDir = getwd(),
  overwrite = FALSE
)
}
\arguments{
\item{TmbFile}{Cpp file name}

\item{CppDir}{directory having the cpp file}

\item{RunDir}{directory for running the code (default: working directory)}

\item{overwrite}{whether the cpp file is overwritten (TRUE) or not (FALSE; default)}
}
\description{
Function for activating cpp file for samuuika with TMB
}
\examples{
\dontrun{
use_est_MSY_tmb()
}

}
