% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlitika.R
\encoding{UTF-8}
\name{visualize_kobe_chart}
\alias{visualize_kobe_chart}
\title{plotting Kobe chart}
\usage{
visualize_kobe_chart(
  assess_data,
  SBrefs = c("SBtarget", "SBlimit", "SBban"),
  Fmsy,
  labeled_year = NULL,
  title = NULL
)
}
\arguments{
\item{assess_data}{assessment data}

\item{SBrefs}{values of SBtarget, SBlimit, and SBban}

\item{Fmsy}{value of Fmsy}

\item{labeled_year}{years to be labeled}

\item{title}{figure title}
}
\description{
plotting Kobe chart
}
