% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messir.R
\encoding{UTF-8}
\name{future_sim}
\alias{future_sim}
\title{future simulation for Surume-Ika}
\usage{
future_sim(
  assess_data,
  Fcurrent,
  multi = 1,
  sim_year = 50,
  nsim = 1000,
  Fcurrent_year = NULL,
  SR = "HS",
  rec_arg = list(a = NULL, b = NULL, sd = 0, rho = 0, resampled_resid = NULL,
    resid_for_bias_correction = NULL),
  bias_correct = TRUE,
  weight,
  num_to_mass_scale = 1/1000,
  M = 0.6,
  Pope = TRUE,
  seed = 12345,
  HCR = FALSE,
  SBrefs = c("SBlimit", "SBban"),
  Ftarget = NULL,
  array_to_tibble = FALSE,
  sim_rec_resid = NULL,
  scenario_name = NULL,
  use_Catch_est = FALSE,
  rec_resid_resample = FALSE,
  backward_resample = FALSE,
  backward_block_duration = 5
)
}
\arguments{
\item{assess_data}{assessment data}

\item{Fcurrent}{current fishing mortality coefficient}

\item{sim_rec_resid}{matrix of sim_year x nsim, DON'T forget to accont for bias-corrected mean}

\item{use_Catch_est}{whether using catch biomass estimated by \code{samuika}}

\item{backward_resample}{whether using backward resampling}

\item{backward_block_duration}{duration of block of backward resampling (default: 5 years)}

\item{use_resid_resample}{nonperametric resampling of \code{rec_arg$resampled resid}}
}
\description{
future simulation for Surume-Ika
}
