% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlitika.R
\encoding{UTF-8}
\name{get_SRdata}
\alias{get_SRdata}
\title{Function to get stock-recruit data from samuika object}
\usage{
get_SRdata(model)
}
\arguments{
\item{model}{samuika object}
}
\description{
Function to get stock-recruit data from samuika object
}
