% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlitika.R
\encoding{UTF-8}
\name{plot_retro_samuika}
\alias{plot_retro_samuika}
\title{plotting retrospective analyses of samuika}
\usage{
plot_retro_samuika(
  samuika_retro,
  mohn = NULL,
  Stock_ID = 0,
  latest_year = NULL,
  max_retro_year = NULL,
  forecast_year = 0,
  is_number = FALSE,
  only_NF = FALSE
)
}
\arguments{
\item{samuika_retro}{samuika_retro object}

\item{is_number}{whether stock and spawning numbers are plotted instead of biomass}

\item{only_NF}{whether plot only N and F (remove S and C)}
}
\description{
plotting retrospective analyses of samuika
}
