% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlitika.R
\encoding{UTF-8}
\name{visualize_future}
\alias{visualize_future}
\title{Function for plotting future_res}
\usage{
visualize_future(
  future_list,
  scenario_name = NULL,
  what_center = "mean",
  CI_range = 0.8,
  title = NULL,
  BRP = NULL,
  line_size = 1,
  legend_position = "right",
  legend_text_size = 1.2,
  legend_key_size = 1,
  strip_text_size = 10,
  base_size = 16,
  BRP_line_size = 1,
  palette_name = "Set1",
  alpha = 0.4
)
}
\arguments{
\item{future_list}{list of \code{future_sim}}

\item{title}{figure title}
}
\description{
Function for plotting future_res
}
