% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messir.R
\encoding{UTF-8}
\name{visualize_future}
\alias{visualize_future}
\title{Function for plotting future_res}
\usage{
visualize_future(
  future_list,
  scenario_name = NULL,
  what_center = "mean",
  CI_range = 0.8,
  title = NULL,
  BRP = NULL
)
}
\arguments{
\item{future_list}{list of \code{future_sim}}

\item{title}{figure title}
}
\description{
Function for plotting future_res
}
