% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlitika.R
\encoding{UTF-8}
\name{visualize_kobe_chart}
\alias{visualize_kobe_chart}
\title{plotting Kobe chart}
\usage{
visualize_kobe_chart(
  assess_data,
  SBrefs = c("SBtarget", "SBlimit", "SBban"),
  Fmsy,
  SBref_label = c("目標基準値案", "限界管理基準値案", "禁漁水準案"),
  use_msy_column = FALSE,
  sep_regime = FALSE,
  labeled_year = NULL,
  title = NULL,
  base_size = 12
)
}
\arguments{
\item{assess_data}{assessment data}

\item{SBrefs}{values of SBtarget, SBlimit, and SBban}

\item{Fmsy}{value of Fmsy}

\item{SBref_label}{labels for SSB reference points}

\item{use_msy_column}{whether SBmsy and Fmsy in 'assess_data' are used (option for different MSY values due to regime shifts)}

\item{sep_regime}{whether using different colours and linetypes between regimes}

\item{labeled_year}{years to be labeled}

\item{title}{figure title}
}
\description{
plotting Kobe chart
}
