% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlitika.R
\encoding{UTF-8}
\name{visualize_yield_curve}
\alias{visualize_yield_curve}
\title{Function for visualizing yield curve from trace_future}
\usage{
visualize_yield_curve(trace_future_res, title = NULL)
}
\arguments{
\item{trace_future_res}{\code{trace_future} object}

\item{title}{figure title}
}
\description{
Function for visualizing yield curve from trace_future
}
